% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart_disease.R
\docType{data}
\name{heart_disease_test_imbalanced_30}
\alias{heart_disease_test_imbalanced_30}
\title{Complementary Test Set for 30\% Imbalanced Training}
\format{
A data frame with 330 rows and 56 columns. The format and variables are identical to \code{\link{heart_disease}}.
}
\usage{
heart_disease_test_imbalanced_30
}
\description{
Test set containing the remaining cases after creating the 30\% imbalanced training set.
Reflects the natural distribution of the original dataset.
Intended for validation of models trained on imbalanced data.
}
\seealso{
\code{\link{heart_disease_train_imbalanced_30}} for the corresponding training set.
}
\keyword{datasets}
